requirements.txt
django
pyjwt
psycopg2
django-cors-headers
passlib
